selections: Remove the "inconsistent" state when not needed
Remove the "inconsistent" state when the user toggles a collection.

According to the docs for GtkToggleButton:

"Normally you would turn off the inconsistent state again if the
user toggles the toggle button. This has to be done manually,
gtk_toggle_button_set_inconsistent() only affects visual appearance,
it doesn’t affect the semantics of the button."

Closes: #3